api-gateway/requirements.txt
Flask==3.1.0
gunicorn==23.0.0
requests==2.32.3
PyJWT==2.10.1
